pyproject.toml
[tool.poetry]
name = "postgres-homeworks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"


[tool.poetry.group.linters.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
mypy = "^1.8.0"
isort = "^5.13.2"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'
warn_unreachable = true
ignore_missing_imports = true

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
